+++ gazebo/gui/QTestFixture.hh
@@ -53,9 +53,10 @@ class QTestFixture : public QObject
   protected: void SetPause(bool _pause);
 
   /// \brief Give the GUI time to process events and trigger main window repaint
-  /// \param[in] _mainWindow Pointer to the main window.
-  /// \param[in] _repeat How many times to repeat the loop.
-  /// \param[in] _ms Time to sleep in milliseconds.
+  /// \param[in] _mainWindow Pointer to the main window. If left void, repaint
+  /// won't be triggered.
+  /// \param[in] _repeat Number of times to repeat the loop.
+  /// \param[in] _ms Time to sleep for each loop in milliseconds.
   protected: void ProcessEventsAndDraw(QMainWindow *_mainWindow = NULL,
       const unsigned int _repeat = 10, const unsigned int _ms = 30);